requirements.txt
flask == 2.3.3
flask-sqlalchemy == 2.5.1
flask-marshmallow == 0.14.0
marshmallow-sqlalchemy == 0.27.0
sqlalchemy == 1.4.30
sqlalchemy-json == 0.5.0
requests == 2.31.0
pyyaml == 6.0.1
flask-migrate == 3.1.0
python-dateutil == 2.8.2
python-dotenv == 0.19.2
gunicorn == 20.1.0
alembic_utils == 0.7.8
pyjwt == 2.6.0
sqlalchemy-utils == 0.39.0
flask-migrate == 3.1.0
python-decouple == 3.5
